requirement.txt
fastapi
uvicorn
sqlalchemy
pgvector
python-dotenv
PyPDF2
pdfminer.six
sentence-transformers
